Merging Sprint 2 Updates

+++ TaskTrackerSolution/TaskTracker/mainLandingPage.cs
@@ -6,6 +6,16 @@ namespace TaskTracker;
 
 class mainLandingPageProcessor
 {
+    private static void DisplayYouCantDoThisMessage(string message)
+    {
+        AnsiConsole.MarkupLine($"Ooops! {message}");
+        userActions.PressEnterToProceed();
+    }
+
+    private static bool CanPerformTaskEntryModificationActions()
+    {
+        return LoginProcessor.CurrentUser.TaskEntries.Count() > 0;
+    }
     private static void displayUseActions()
     {
         var userSelectedAction = AnsiConsole.Prompt(
@@ -15,10 +25,11 @@ class mainLandingPageProcessor
                 .MoreChoicesText("[grey](Move up and down to reveal more options)[/]")
                 .AddChoices(new []{
                     "Add Task",
-                    "View Task Detail",
-                    "Update Task",
                     "Delete Task",
                     "Purge Completed Tasks",
+                    "Sort Tasks View",
+                    "Update Task",
+                    "View Task Detail",
                     "Log out",
                     "Quit"
                 })
@@ -26,23 +37,54 @@ class mainLandingPageProcessor
 
         switch(userSelectedAction)
         {
+            case "Sort Tasks View":
+                if (CanPerformTaskEntryModificationActions())
+                    userActions.SortTaskView();
+                else
+                {
+                    DisplayYouCantDoThisMessage("You haven't registered any tasks.  Please start with the 'Add Task' action.");
+                    DisplayMainLandingPage();
+                }
+                break;
             case "Add Task":
                 userActions.AddTask();
                 break;
             case "View Task Detail":
-                userActions.ViewTaskDetail();
+                    
+                if (CanPerformTaskEntryModificationActions())
+                    userActions.ViewTaskDetail();
+                else
+                {
+                    DisplayYouCantDoThisMessage("You haven't registered any tasks.  Please start with the 'Add Task' action.");
+                    DisplayMainLandingPage();
+                }
                 break;
-            case "Update Task":
-                userActions.UpdateTask();
+            case "Update Task":                    
+                if (CanPerformTaskEntryModificationActions())
+                    userActions.UpdateTask();
+                else
+                {
+                    DisplayYouCantDoThisMessage("You haven't registered any tasks.  Please start with the 'Add Task' action.");
+                    DisplayMainLandingPage();
+                }
                 break;
-            case "Delete Task":
-                userActions.DeleteTask();
+            case "Delete Task":                    
+                if (CanPerformTaskEntryModificationActions())
+                    userActions.DeleteTask();
+                else
+                {
+                    DisplayYouCantDoThisMessage("You haven't registered any tasks.  Please start with the 'Add Task' action.");
+                    DisplayMainLandingPage();
+                }
                 break;
-            case "Purage Completed Tasks":
+            case "Purge Completed Tasks":
                 userActions.PurgeTasks();
                 break;
             case "Log out":
-                userActions.LogoutAction();
+                //clear the screen
+                AnsiConsole.Clear();
+                //go to the login page
+                Program.showWelcomePage();
                 break;
             case "Quit":
                 userActions.QuitAction();
@@ -58,7 +100,7 @@ class mainLandingPageProcessor
             throw new NullReferenceException("User Model is null.");
 
         AnsiConsole.Clear();
-        var rule = new Rule("Welcome To the Main Page");
+        var rule = new Rule("Trask Tracker Main Page: Task Listing");
         rule.RuleStyle("gold3 bold");
         AnsiConsole.Write(rule);
         AnsiConsole.MarkupLine($"User: [gold3 bold]{LoginProcessor.CurrentUser.UserName}[/]");
@@ -77,7 +119,7 @@ class mainLandingPageProcessor
         else
         {
             AnsiConsole.WriteLine("");
-            AnsiConsole.MarkupLine("No tasks defined.  Choose an option from below to get started :backhand_index_pointing_down:");
+            AnsiConsole.MarkupLine("No tasks registered yet.  Choose an option from below to get started :backhand_index_pointing_down:");
             AnsiConsole.WriteLine("");
         }
         //show user options
